refactor: make common log config into a shared struct
