+++ governance/Cargo.toml
@@ -15,7 +15,6 @@ path = "src/main.rs"
 [dependencies]
 tokio.workspace = true 
 tracing.workspace = true
-tracing-subscriber.workspace = true
 chrono.workspace = true 
 clap.workspace = true 
 anyhow.workspace = true
@@ -27,7 +26,6 @@ futures.workspace = true
 deadpool-diesel.workspace = true
 diesel.workspace = true
 orm.workspace = true
-clap-verbosity-flag.workspace = true
 tokio-retry.workspace = true
 
 [build-dependencies]
