+++ parameters/src/config.rs
@@ -1,7 +1,7 @@
 use core::fmt;
 use std::fmt::Display;
 
-use clap_verbosity_flag::{InfoLevel, Verbosity};
+use shared::log_config::LogConfig;
 
 #[derive(clap::ValueEnum, Clone, Debug, Copy)]
 pub enum CargoEnv {
@@ -15,18 +15,6 @@ impl Display for CargoEnv {
     }
 }
 
-#[derive(clap::ValueEnum, Clone, Debug, Copy)]
-pub enum LogFormat {
-    Json,
-    Text,
-}
-
-impl Display for LogFormat {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", format!("{:?}", self).to_lowercase())
-    }
-}
-
 #[derive(clap::Parser)]
 pub struct AppConfig {
     #[clap(long, env)]
@@ -35,12 +23,9 @@ pub struct AppConfig {
     #[clap(long, env)]
     pub database_url: String,
 
-    #[command(flatten)]
-    pub verbosity: Verbosity<InfoLevel>,
-
-    #[clap(long, env, default_value_t = LogFormat::Text)]
-    pub log_format: LogFormat,
-
     #[clap(long, env, default_value_t = 30)]
     pub sleep_for: u64,
+
+    #[clap(flatten)]
+    pub log: LogConfig,
 }
