+++ parameters/src/main.rs
@@ -4,7 +4,6 @@ use std::time::Duration;
 
 use chrono::NaiveDateTime;
 use clap::Parser;
-use clap_verbosity_flag::LevelFilter;
 use deadpool_diesel::postgres::Object;
 use namada_sdk::state::EPOCH_SWITCH_BLOCKS_DELAY;
 use namada_sdk::time::{DateTimeUtc, Utc};
@@ -12,7 +11,7 @@ use orm::gas::GasPriceDb;
 use orm::migrations::run_migrations;
 use orm::parameters::ParametersInsertDb;
 use parameters::app_state::AppState;
-use parameters::config::{AppConfig, LogFormat};
+use parameters::config::AppConfig;
 use parameters::repository;
 use parameters::services::{
     namada as namada_service, tendermint as tendermint_service,
@@ -23,29 +22,12 @@ use shared::error::{AsDbError, AsRpcError, ContextDbInteractError, MainError};
 use tendermint_rpc::HttpClient;
 use tokio::sync::{Mutex, MutexGuard};
 use tokio::time::Instant;
-use tracing::Level;
-use tracing_subscriber::FmtSubscriber;
 
 #[tokio::main]
 async fn main() -> Result<(), MainError> {
     let config = AppConfig::parse();
 
-    let log_level = match config.verbosity.log_level_filter() {
-        LevelFilter::Off => None,
-        LevelFilter::Error => Some(Level::ERROR),
-        LevelFilter::Warn => Some(Level::WARN),
-        LevelFilter::Info => Some(Level::INFO),
-        LevelFilter::Debug => Some(Level::DEBUG),
-        LevelFilter::Trace => Some(Level::TRACE),
-    };
-    if let Some(log_level) = log_level {
-        let subscriber = FmtSubscriber::builder().with_max_level(log_level);
-
-        match config.log_format {
-            LogFormat::Text => subscriber.init(),
-            LogFormat::Json => subscriber.json().flatten_event(true).init(),
-        };
-    }
+    config.log.init();
 
     let client =
         Arc::new(HttpClient::new(config.tendermint_url.as_str()).unwrap());
