+++ seeder/src/config.rs
@@ -1,7 +1,7 @@
 use core::fmt;
 use std::fmt::Display;
 
-use clap_verbosity_flag::{InfoLevel, Verbosity};
+use shared::log_config::LogConfig;
 
 #[derive(clap::ValueEnum, Clone, Debug, Copy)]
 pub enum CargoEnv {
@@ -12,18 +12,6 @@ pub enum CargoEnv {
 impl Display for CargoEnv {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{:?}", self)
-    }
-}
-
-#[derive(clap::ValueEnum, Clone, Debug, Copy)]
-pub enum LogFormat {
-    Json,
-    Text,
-}
-
-impl Display for LogFormat {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", format!("{:?}", self).to_lowercase())
     }
 }
 
@@ -53,9 +41,6 @@ pub struct AppConfig {
     #[clap(long, env, default_value_t = 10)]
     pub total_balances: u64,
 
-    #[command(flatten)]
-    pub verbosity: Verbosity<InfoLevel>,
-
-    #[clap(long, env, default_value_t = LogFormat::Text)]
-    pub log_format: LogFormat,
+    #[clap(flatten)]
+    pub log: LogConfig,
 }
