+++ shared/Cargo.toml
@@ -17,6 +17,8 @@ anyhow.workspace = true
 async-stream.workspace = true
 bimap.workspace = true
 bigdecimal.workspace = true
+clap.workspace = true
+clap-verbosity-flag.workspace = true
 futures-core.workspace = true
 futures-util.workspace = true
 futures.workspace = true
@@ -35,5 +37,6 @@ thiserror.workspace = true
 tokio-retry.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+tracing-subscriber.workspace = true
 fake.workspace = true
 rand.workspace = true
