+++ shared/src/lib.rs
@@ -10,6 +10,7 @@ pub mod gas;
 pub mod genesis;
 pub mod header;
 pub mod id;
+pub mod log_config;
 pub mod parameters;
 pub mod proposal;
 pub mod public_key;
